Create Number_guessing_game.py
Number guessing game! 
The game picks a number between 1 and 10, and you have to guess what it is. It will give you hints as you go along (whether the secret number is higher or lower than your guess.) For the smart cookies, it's possible to win every time, if you work out the method.